requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
openai==0.27.0
python-dotenv==1.0.0
pytest==7.4.3